refactor: implement depth recurrence, SwiGLU, and 32k vocab support in train_gpt.py while removing train_gpt_recurrent.py.
